requirements.txt
discord
PyNaCl==1.3.0
pandas
dnspython==1.16.0
requests
Pillow
scikit-image
opencv-python-headless
pytube3
pydub
glitch_this
tweepy
bs4